+++ NPS Browser/db/RealmStorageContext.swift
@@ -71,7 +71,7 @@ extension RealmStorageContext {
         }
         
         try self.safeWrite {
-            let newObject = realm.create(model as! Object.Type, value: [], update: false) as! T
+            let newObject = realm.create(model as! Object.Type, value: [], update: .error) as! T
             completion(newObject)
         }
     }